refactor(bot): Moved code to own files
